+++ App.js
@@ -3,19 +3,8 @@ import React, {useEffect, useState} from 'react';
 import auth from '@react-native-firebase/auth';
 
 import 'react-native-gesture-handler';
-import {NavigationContainer} from '@react-navigation/native';
-import {createDrawerNavigator} from '@react-navigation/drawer';
-import {DrawerContent} from "./screens/Drawer";
-import AboutScreen from "./screens/AboutScreen/aboutScreen";
-import SettingsScreen from "./screens/SettingsScreen";
-
-import RootStackScreen from "./screens/RootStackScreen"
-import EventsScreen from './screens/EventsScreen';
-import ProfileScreen from './screens/ProfileScreen';
-import NewEventScreen from "./screens/NewEventScreen";
-import HomeScreen from "./screens/HomeScreen";
-
-const Drawer = createDrawerNavigator();
+import RootStackScreen from "./navigation/RootStackScreen"
+import RootUserStackScreen from "./navigation/RootUserStackScreen";
 
 export default function App() {
 
@@ -33,22 +22,11 @@ export default function App() {
 
     if (!user) {
         return (
-            <NavigationContainer>
-                <RootStackScreen/>
-            </NavigationContainer>
+            <RootStackScreen/>
         );
     }
 
     return (
-        <NavigationContainer>
-            <Drawer.Navigator drawerContent={props => <DrawerContent {...props} />}>
-                <Drawer.Screen name="HomeScreen" component={HomeScreen}/>
-                <Drawer.Screen name="EventsScreen" component={EventsScreen}/>
-                <Drawer.Screen name="ProfileScreen" component={ProfileScreen}/>
-                <Drawer.Screen name="AboutScreen" component={AboutScreen}/>
-                <Drawer.Screen name="SettingsScreen" component={SettingsScreen}/>
-                <Drawer.Screen name="NewEventScreen" component={NewEventScreen}/>
-            </Drawer.Navigator>
-        </NavigationContainer>
+        <RootUserStackScreen/>
     );
 }
