+++ screens/HomeScreen/index.js
@@ -5,13 +5,13 @@ import HomeSearch from "../../components/HomeSearch";
 import Ionicons from "react-native-vector-icons/Ionicons";
 import styles from "./styles";
 
-const HomeScreen = ({navigation}) => {
+const HomeScreen = (props) => {
     return (
         <View>
             <StatusBar translucent backgroundColor="transparent" barStyle="dark-content"/>
             <HomeMap/>
             <View style={styles.headerBar}>
-                <Ionicons size={32} name='menu' onPress={() => navigation.openDrawer()}/>
+                <Ionicons size={32} name='menu' onPress={() => props.navigation.openDrawer()}/>
             </View>
             <HomeSearch/>
         </View>
